Reorganized and greatly added to Messaging. Checkpoint - not stable.
