--- Src/RCommon.ApplicationServices.Messaging/IDistributedEventHandler.cs
@@ -1,16 +1,0 @@
-﻿
-using MassTransit;
-using System.Threading.Tasks;
-
-namespace RCommon.ApplicationServices.Messaging
-{
-    public interface IDistributedEventHandler<in TDistributedEvent> : IDistributedEventHandler, IConsumer<TDistributedEvent>
-        where TDistributedEvent : DistributedEvent
-    {
-        
-    }
-
-    public interface IDistributedEventHandler
-    {
-    }
-}
