+++ Src/RCommon.ApplicationServices/RCommon.ApplicationServices.csproj
@@ -18,7 +18,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\RCommon.DataServices\RCommon.DataServices.csproj" />
-    <ProjectReference Include="..\RCommon.Validation\RCommon.Validation.csproj" />
     <ProjectReference Include="..\RCommon.Core\RCommon.Core.csproj" />
   </ItemGroup>
