+++ Src/RCommon.Messaging.JustSaying/RCommon.Messaging.JustSaying.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="JustSaying" Version="7.1.1" />
+  </ItemGroup>
+
+</Project>
