+++ Src/RCommon.Messaging.MassTransit/IMassTransitEventHandler.cs
@@ -1,0 +1,14 @@
+﻿using MassTransit;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RCommon.Messaging.MassTransit
+{
+    public interface IMassTransitEventHandler<in TDistributedEvent> : IDistributedEventHandler, IConsumer<TDistributedEvent>
+        where TDistributedEvent : DistributedEvent
+    {
+    }
+}
