+++ Src/RCommon.Messaging/ApplicationMessagingConfigurationExtensions.cs
@@ -1,4 +1,4 @@
-﻿using MassTransit;
+﻿
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using RCommon.ApplicationServices.Messaging;
@@ -12,15 +12,7 @@ using System.Threading.Tasks;
 namespace RCommon
 {
     public static class ApplicationMessagingConfigurationExtensions
-    {
-
-        public static IRCommonConfiguration WithMassTransit(this IRCommonConfiguration config, Action<IBusRegistrationConfigurator> massTransitConfig)
-        {
-            config.Services.AddScoped<IDistributedEventBroker, DistributedEventBroker>();
-            config.Services.AddMassTransit(massTransitConfig);
-            return config;
-        }
-
+    { 
         public static IRCommonConfiguration AddDisributedUnitOfWorkToMediatorPipeline(this IRCommonConfiguration config)
         {
             config.Services.AddTransient(typeof(IPipelineBehavior<,>), typeof(DistributedUnitOfWorkBehavior<,>));
