+++ Src/RCommon.Messaging/IDynamicDistributedEventHandler.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace RCommon.ApplicationServices.Messaging
+namespace RCommon.Messaging
 {
     public interface IDynamicDistributedEventHandler
     {
