+++ Src/RCommon.Messaging/RCommon.Messaging.csproj
@@ -25,10 +25,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="MassTransit" Version="8.1.1" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\RCommon.ApplicationServices\RCommon.ApplicationServices.csproj" />
   </ItemGroup>
