--- Src/RCommon.Validation/IEntityValidator.cs
@@ -1,33 +1,0 @@
-﻿#region license
-//Copyright 2010 Ritesh Rao 
-
-//Licensed under the Apache License, Version 2.0 (the "License"); 
-//you may not use this file except in compliance with the License. 
-//You may obtain a copy of the License at 
-
-//http://www.apache.org/licenses/LICENSE-2.0 
-
-//Unless required by applicable law or agreed to in writing, software 
-//distributed under the License is distributed on an "AS IS" BASIS, 
-//WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
-//See the License for the specific language governing permissions and 
-//limitations under the License. 
-#endregion
-
-namespace RCommon.Validation
-{
-    /// <summary>
-    /// Interface implemented by different flavors of validators that provide validation
-    /// logic on domain entities.
-    /// </summary>
-    /// <typeparam name="TEntity"></typeparam>
-    public interface IEntityValidator<TEntity>
-    {
-        /// <summary>
-        /// Validates an entity against all validations defined for the entity.
-        /// </summary>
-        /// <param name="entity">The <typeparamref name="TEntity"/> to validate.</param>
-        /// <returns>A <see cref="ValidationResult"/> that contains the results of the validation.</returns>
-        ValidationResult Validate(TEntity entity);
-    }
-}
