--- Src/RCommon.Validation/RCommon.Validation.csproj
@@ -1,34 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageIcon>rcommon-icon.png</PackageIcon>
-    <PackageIconUrl />
-    <PackageProjectUrl>https://github.com/Reactor2Team/RCommon</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/Reactor2Team/RCommon</RepositoryUrl>
-    <Description>A cohesive set of .NET 7 infrastructure libraries that utilizes abstractions for persistence, unit of work/transactions, distributed events, distributed transactions, and more.</Description>
-    <Authors>Reactor2, Jason Webb</Authors>
-    <Company>RCommon</Company>
-    <Version>1.0.2.1</Version>
-    <RepositoryType>GitHub</RepositoryType>
-    <PackageTags>RCommon, object validation, validation abstractions, c#, .NET</PackageTags>
-    <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\RCommon.Core\RCommon.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\RCommon.Core\rcommon-icon.png">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="FluentValidation" Version="11.8.0" />
-  </ItemGroup>
-
-</Project>
